+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/simple_table/simple_table_widgets/simple_table_add_column_button.dart
@@ -74,7 +74,7 @@ class _SimpleTableAddColumnHoverButtonState
                   hoveringTableCell.columnLength;
             }
             return Positioned(
-              top: SimpleTableConstants.tableTopPadding -
+              top: SimpleTableConstants.tableHitTestTopPadding -
                   SimpleTableConstants.cellBorderWidth,
               bottom: SimpleTableConstants.addColumnButtonBottomPadding,
               right: 0,
