+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/simple_table/simple_table_widgets/widgets.dart
@@ -1,3 +1,4 @@
+export 'simple_table_action_sheet.dart';
 export 'simple_table_add_column_and_row_button.dart';
 export 'simple_table_add_column_button.dart';
 export 'simple_table_add_row_button.dart';
@@ -7,5 +8,6 @@ export 'simple_table_basic_button.dart';
 export 'simple_table_border_builder.dart';
 export 'simple_table_column_resize_handle.dart';
 export 'simple_table_divider.dart';
+export 'simple_table_more_action_popup.dart';
 export 'simple_table_reorder_button.dart';
 export 'simple_table_widget.dart';
